{"schema":"decodex/commit/1","summary":"Modularize recovery and harness status boundaries","authority":"manual"}
