+++ apps/decodex/src/orchestrator/execution_failure/review_handoff_drift/recovery/transition.rs
@@ -1,0 +1,33 @@
+use color_eyre::eyre;
+
+use crate::orchestrator::execution_failure::{
+	IssueRunPlan, Result, WorkflowDocument,
+	review_handoff_drift::types::ReviewHandoffStateDriftTransition,
+};
+
+pub(in crate::orchestrator::execution_failure::review_handoff_drift) fn review_handoff_state_drift_success_transition(
+	workflow: &WorkflowDocument,
+	issue_run: &IssueRunPlan,
+) -> Result<Option<ReviewHandoffStateDriftTransition>> {
+	let tracker_policy = workflow.frontmatter().tracker();
+	let success_state = tracker_policy.success_state();
+	let current_state = issue_run.issue.state.name.as_str();
+
+	if current_state == success_state {
+		return Ok(Some(ReviewHandoffStateDriftTransition::AlreadySuccess));
+	}
+	if current_state != tracker_policy.in_progress_state()
+		&& current_state != tracker_policy.failure_state()
+	{
+		return Ok(None);
+	}
+
+	let state_id = issue_run.issue.state_id_for_name(success_state).ok_or_else(|| {
+		eyre::eyre!(
+			"State `{success_state}` was not found for issue `{}` during review handoff state drift recovery.",
+			issue_run.issue.identifier
+		)
+	})?;
+
+	Ok(Some(ReviewHandoffStateDriftTransition::MoveToSuccess(state_id.to_owned())))
+}
